allow to run script from this repo

+++ tbotlib.sh
@@ -14,20 +14,23 @@ tmp_folder=$(mktemp -d)
 check_new_version=$(curl -sS ${LIB_RAW_URL} | grep -m1 VERSION | cut -d':' -f2)
 current_version=$(cat ${BASEDIR}/tbotlib.sh | grep -m1 VERSION | cut -d':' -f2)
 if [[ "${current_version}" != "${check_new_version}" ]]; then
-    echo "[WARN] tbotlib is out of date, we recomend to get new version."
+    echo "[WARN] tbotlib is out of date, we recomend to get new version ${check_new_version}"
     echo "[INFO] tbotlib script: ${LIB_RAW_URL}"
 else
     echo "[INFO] tbotlib is up to date!"
 fi
 
-if [[ ! -d ${BASEDIR}/tbotlibs ]]; then
+if [[ ! -d ${BASEDIR}/tbotlibs ]] && \
+        [[ "$(git config --get remote.origin.url)" != "${LIB_REPO}" ]]; then
     git clone --quiet --single-branch --branch ${LIB_BRANCH} ${LIB_REPO} ${tmp_folder} > /dev/null
     cp -r ${tmp_folder}/tbotlibs ${LIBS_FOLDER}
     rm -fr ${tmp_folder}    
 fi
 
-[[ $(cat ${BASEDIR}/.gitignore | grep tbotlibs) ]] || \
+if [[ ! $(cat ${BASEDIR}/.gitignore | grep tbotlibs) ]] && \
+        [[ "$(git config --get remote.origin.url)" != "${LIB_REPO}" ]]; then
     echo -e "\n\n#Telegram bot Libs\ntbotlibs\n-" >> ${BASEDIR}/.gitignore
+fi
 
 libs_list=($(find ${BASEDIR}/tbotlibs -name "*.sh"))
 for f in ${libs_list[@]}; do
